+++ sitegen/site.py
@@ -1,15 +1,18 @@
 from __future__ import annotations
 import logging
 import os
-from datetime import datetime, timezone
+import pickle
+import tempfile
+import hashlib
+from datetime import datetime, timezone, timedelta
 from enum import Enum
 from pathlib import Path
-from typing import Final, List, Union, TypeAlias, Any, Type
+from typing import Final, List, Union, TypeAlias, Any, Type, Optional
 from collections.abc import Generator, Callable
 from jinja2 import Environment, FileSystemLoader, Template
 from .context import FileType, BuildContext, Metrics
 from .templates import RSS_FALLBACK, SITEMAP_FALLBACK
-from .build import MarkdownPage, DEFAULT_EXTENSIONS
+from .build import Page, MarkdownPage, DEFAULT_EXTENSIONS
 from .exec import FileTypeError
 
 logger = logging.getLogger(__name__)
@@ -17,10 +20,10 @@ logger = logging.getLogger(__name__)
 SOURCE_DIR: Final[Path] = Path("source")
 DEST_DIR: Final[Path] = Path("build")
 TEMPLATE_DIR: Final[Path] = Path("templates")
+OBJ_CACHE_DIR: Final[Path] = Path(".cache")
 URL_BASE: Final[str] = "https://itsdanjc.com"
 URL_INDEX: Final[str] = "index.html"
 
-Page: TypeAlias = Union[MarkdownPage]
 TreeItem: TypeAlias = Union["TreeNode", Page]
 
 
@@ -35,11 +38,11 @@ class SortKey(Enum):
 
 class TreeNode:
     path: Final[Path]
-    parent: Final[TreeNode | None]
+    parent: Final[Optional[TreeNode]]
     pages: list[Page]
     sub_dirs: list[TreeNode]
 
-    def __init__(self, path: Path, parent: TreeNode | None = None):
+    def __init__(self, path: Path, parent: Optional[TreeNode] = None):
         self.path = path
         self.parent = parent
         self.pages = []
@@ -55,10 +58,10 @@ class TreeNode:
 
     def __contains__(self, item: TreeItem) -> bool:
         if isinstance(item, Page):
-            return any(i for i in self)
+            return any(i == item for i in self)
 
         if isinstance(item, TreeNode):
-            return any(i for i in self.walk())
+            return any(i == item for i in self.walk())
 
         return False
 
@@ -124,11 +127,35 @@ class TreeBuilder:
     node_not_root: bool
     metrics: Metrics
 
-    def __init__(self, site: SiteRoot):
+    # Used for cache file.
+    # If a potential risk for breaking changes with builder, or any tree objects, this value
+    # should be changed.
+    cache_version = 1
+
+    def __init__(self, site: SiteRoot, cache_delta: timedelta):
+        """
+        Perform an index of the source files.
+
+        Added in v0.3.0: Indexing now uses a cache to make repeat runs faster.
+        If this command is run again shortly after a previous one, the cached
+        index will be reused.
+
+        :param site: the site to index.
+        :param cache_delta: the duration after which the cache expires.
+        """
+
         self.site = site
         self.node = site.tree
+        self.cache_file = self.site.root.joinpath(OBJ_CACHE_DIR, f"srctree.bin")
+        self.cache_hash_file = self.site.root.joinpath(OBJ_CACHE_DIR, f"srctree_h.bin")
 
         with Metrics() as self.metrics:
+            cache = self.cache(cache_delta)
+            if cache:
+                self.site.tree = cache
+                return
+
+            logger.info("Starting index...")
             for curr_dir, dir_list, file_list in os.walk(self.site.source_dir):
                 self.node_path = Path(curr_dir)
                 self.node_dir_list = dir_list
@@ -142,7 +169,14 @@ class TreeBuilder:
                 self.create_directory_nodes()
                 self.create_file_nodes()
 
+            logger.debug("Saving cache %s", self.cache_file)
+            self.write_cache()
+
     def create_directory_nodes(self):
+        """
+        Iterate through directory nodes, create a dir object,
+        and add to sub_dirs list.
+        """
         for sub_dir in self.node_dir_list:
             node_path = Path(self.node_path, sub_dir)
             self.node.sub_dirs.append(
@@ -150,6 +184,11 @@ class TreeBuilder:
             )
 
     def create_file_nodes(self):
+        """
+        Iterate through file nodes, create a page object,
+        and add to pages list.
+        """
+
         for file in self.node_file_list:
             file_path = Path(self.node_path, file).relative_to(
                 self.site.source_dir
@@ -170,6 +209,99 @@ class TreeBuilder:
             self.node.pages.append(
                 PageNode.create(context, self.site.env)
             )
+
+    def cache(self, delta: timedelta) -> Optional[TreeNode]:
+        """
+        Wrapper for self.read_cache() - logs exceptions.
+        :return: TreeNode, or None if a reindex is required.
+        """
+        cache_last_mod = datetime.fromtimestamp(
+            self.cache_file.stat().st_mtime
+            if self.cache_file.exists() else 0
+        )
+
+        try:
+            expired = (datetime.now() - cache_last_mod) > delta
+            if expired:
+                logger.debug("Cache expired, will reindex")
+                return None
+
+            f = self.read_cache()
+            logger.debug("Skipping index, using cache")
+            return f
+
+        except FileNotFoundError:
+            logger.debug("Missing cache files, will reindex")
+
+        except ValueError:
+            logger.warning("Cache file invalid, will reindex")
+
+        except OSError as e:
+            logger.exception("Failed to read cache, will reindex", exc_info=e)
+
+    def write_cache(self) -> None:
+        """
+        Pickle the site tree, and write it to disk.
+
+        This is very useful as a performance boost for larger
+        sites to avoid needing to reindex.
+        """
+        tree = self.site.tree
+        cache_dir = self.cache_file.parent
+        cache_dir.mkdir(exist_ok=True)
+
+        payload = {"version": self.cache_version, "tree": tree}
+        dump = pickle.dumps(payload)
+        hash_str = hashlib.sha256(dump).hexdigest()
+
+        # Write dump to temp location
+        with tempfile.NamedTemporaryFile("wb", delete=False, dir=cache_dir) as fout:
+            fout.write(dump)
+            temp_pickle = fout.name
+
+        # Write hash to temp location
+        with tempfile.NamedTemporaryFile("w", delete=False, dir=cache_dir) as fout:
+            fout.write(hash_str)
+            temp_hash = fout.name
+
+        # Replace temp files, with actual locations.
+        os.replace(temp_pickle, self.cache_file)
+        os.replace(temp_hash, self.cache_hash_file)
+
+    def read_cache(self) -> TreeNode:
+        """
+        Read cache artifact from disk.
+
+        This is very useful as a performance boost for larger
+        sites to avoid needing to reindex.
+
+        :return: Root tree node.
+
+        :raises FileNotFoundError: if any of the cache files are missing.
+        :raises ValueError: a mismatch with cache version or hash.
+        :raises OSError: other reason preventing read.
+        """
+        if not (self.cache_file.exists() and self.cache_hash_file.exists()):
+            raise FileNotFoundError("Cache file not found")
+
+        try:
+            read = self.cache_file.read_bytes()
+            tree_hash = hashlib.sha256(read).hexdigest()
+            data = pickle.loads(read)
+
+            hash_not_match = (tree_hash != self.cache_hash_file.read_text())
+            version_not_match = (data["version"] != self.cache_version)
+
+            if hash_not_match:
+                raise ValueError("Cache file mismatch")
+
+            if version_not_match:
+                raise ValueError("Cache version mismatch")
+
+            return data["tree"]
+
+        except (pickle.UnpicklingError, EOFError, OSError) as e:
+            raise OSError("Cache read error") from e
 
 
 class SiteRoot: