+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 ## Fixes
 - When creating rss and sitemap, it no longer needs to rerender all pages again.
 
+## Added
+- Caching file index, gives huge performance boost for large sites.
+
 ## 0.2.0
 ### Added
 - RSS Feed.